fix test loadbal after rebase
